Use unsigned int for ackId
signed integer overflow is undefined behaviour, and WRandom::get()
actually produces an unsigned integer. That random integer may be close
enough to INT_MAX or INT_MIN that it actually matters, and overflow is
possible.

Also simplified the expectedAckId_ - updateId < 5 check. We can just
rely on the well defined unsigned integer overflow behaviour. It's not
necessary to check whether updateId < expectedAckId_.

Added operator<< (unsigned int) to WStringStream so we don't need to
cast it to long long.

+++ src/web/WebSession.C
@@ -1943,7 +1943,7 @@ void WebSession::handleWebSocketMessage(std::weak_ptr<WebSession> session,
 	      lock->asyncResponse_ = nullptr;
 	    }
 
-	    lock->renderer_.ackUpdate(Utils::stoi(*connectedE));
+	    lock->renderer_.ackUpdate(static_cast<unsigned int>(Utils::stoul(*connectedE)));
 	    lock->webSocketConnected_ = true;
 	  }
 
@@ -2443,7 +2443,7 @@ void WebSession::notify(const WEvent& event)
 	  WebRenderer::AckState ackState = WebRenderer::CorrectAck;
 	  if (invalidAckId && ackIdE) {
 	    try {
-	      ackState = renderer_.ackUpdate(Utils::stoi(*ackIdE));
+	      ackState = renderer_.ackUpdate(static_cast<unsigned int>(Utils::stoul(*ackIdE)));
 	      if (ackState != WebRenderer::BadAck)
 		invalidAckId = false;
 	    } catch (const std::exception& e) {
